fix: realiza los cambios solicitados por el cliente.

+++ src/main/java/com/albertogalvez/Kinalapp/service/IVentasService.java
@@ -6,10 +6,10 @@ import java.util.Optional;
 
 public interface IVentasService {
     List<Ventas> listarTodos();
-    Optional<Ventas> buscarPorId(Integer id);
+    Optional<Ventas> buscarPorId(Long id);
     Ventas guardar(Ventas ventas);
-    Ventas actualizar(Integer id, Ventas ventas);
-    void eliminar(Integer id);
-    boolean existePorId(Integer id);
+    Ventas actualizar(Long id, Ventas ventas);
+    void eliminar(Long id);
+    boolean existePorId(Long id);
     List<Ventas> listarPorEstado(int estado);
 }
